Refactor flowcell to models module
